+++ Assignment1/src/should_live.c
@@ -11,7 +11,7 @@
     &maxVal: the cell should live
     0:       the cell should die
 */
-unsigned char should_live(unsigned const int k, unsigned const int i, unsigned const char *world, unsigned const int *maxVal)
+unsigned char should_live(unsigned const int k,  unsigned const long i, unsigned const char *world, unsigned const int *maxVal)
 {
     int result = 0;   // char is not enough to store the first sum
     register unsigned const int row = i/k;
